Move activity processor code to the correct place.
